remove long reference

+++ selfdrive/car/car_helpers.py
@@ -106,7 +106,7 @@ def fingerprint(logcan, sendcan, has_relay):
 
   if cached_fingerprint is not None and use_car_caching:  # if we previously identified a car and fingerprint and user hasn't disabled caching
     cached_fingerprint = json.loads(cached_fingerprint)
-    finger[0] = {long(key): value for key, value in cached_fingerprint[1].items()}  # not sure if dict of longs is required
+    finger[0] = {key: value for key, value in cached_fingerprint[1].items()}
     return (str(cached_fingerprint[0]), finger, vin)
